Navigation and hero page working
Added in proper navigation to the pages I'm using to build the database. Also added a hero to the home page for a bang on entrance. Search bar may need to remove the search button  in order to be more appealing. Working on the bottom two buttons being squared up to work better with a touch screen.