added in shopping cart functionality to context
As a note to future me, in case I forget, context is more or less passing a prop throughout the application as I have it set now. In app file, the entire app is wrapped with the shopping cart context. In the context file, I've added functionality to add, remove, and modify items in the cart to handle the food portion of the app.