Added authentication
Added in authentication in order to change the rules for my Firebase to allow changes once more. Firebase auth now works with google sign-in through clicking the magnifying glass. This obviously needs to change at some point, but so long as it works, I can now fix the rules for Firebase.