further updates to firestore use
Updated how firestore is manipulated to add and remove items from cart. App does boot up, but there's currently an error when you attempt to add an item. There's try catch now so the app won't crash but it's not currently working. Will need to test how the sign-in works to make sure that's not causing the issue.