CSS cleanup, changes to routing, new API call
Probably should have saved sooner. Started with cleaning up the shopping cart, it will now correctly display items in a familiar format with the total at bottom. The Checkout button just clears cart as of now and informs user to pick up at counter. Cleaned up the CSS being used to display the game cards so they're (almost) uniform across cards. Then fixed the routing so that it works from the game list. Removed location state as making a call to firebase was easier. Finally, redid the css on the game page so that it displays in a relevant manner with everything legible and important information easy to find.