food area updated
changed up the foodHome component and the foodCard component to match up with the other area's designs. I swapped out the image in a background for an abstract svg to make it look fun but more modern. It's still probably a bit too rich but we'll come back to that later.